app/requirements.txt
fastapi==0.85.1
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==0.21.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
python-multipart
python-jose